Implement label parsing as prefix for 1pass
This label parsing is considered a part of the
first pass, but must be done explicitly before hand.
This allows for labels to appear in code before they are
declared.  The way the software is currently implemented,
unknown labels will be dealt with in either the second phase
of the first pass or the second pass, depending on where they
lie.