Added the loader, should be fully functional
Decided to use a dict here =) I am excited to see how this could
turn out.  Basically, we're using the high level structure so
that we don't need to contain the entire contents of memory,
which could save a little bit of memory for smaller programs.
Moreover, it allows us to use exceptions to detect when to stop
the VM (via a KeyError)